Split CheckVersionIncrement helpers into focused files
- Move the `version.gradle.kts` reading/git helpers (`VERSION_FILE`, `GitResult`,
  `headVersionFile`, `baseVersionFile`, `gitShow`) into `VersionGradleFile.kt`,
  raising the cross-file ones to `internal`.
- Extract `MavenMetadata` and `Versioning` into `MavenMetadata.kt` (now `internal`),
  document why `Versioning.versions` is `var`, and convert `fetchAndParse`'s KDoc
  `<p>` to Markdown.
- `CheckVersionIncrement.kt` keeps only the task and drops its now-unused imports.
- Add `MavenMetadataSpec`, a Jackson round-trip that guards the `var`/`internal`
  setters — a `val` would silently empty the parsed version list.

No behavior change.
